hoàn thiện chức năng trang admin

+++ backend/src/main/java/com/cdweb/laptopStore/entities/OrderItem.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -34,5 +35,5 @@ public class OrderItem {
     @Column(nullable = false)
     private Integer quantity;
 
-    private Double itemPrice;
+    private BigDecimal itemPrice;
 }
